chore: Move tests to root.
