--- src/tests/Usings.cs
@@ -1,2 +1,0 @@
-global using Microsoft.OpenApi.ApiManifest;
-global using Xunit;
