+++ tests/ApiManifest.Tests.csproj
@@ -1,10 +1,11 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
+	<AssemblyName>Microsoft.OpenApi.ApiManifest.Tests</AssemblyName>
+	<RootNamespace>Microsoft.OpenApi.ApiManifest.Tests</RootNamespace>
     <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
   </PropertyGroup>
@@ -23,7 +24,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\lib\apimanifest.csproj" />
+    <ProjectReference Include="..\src\lib\apimanifest.csproj" />
   </ItemGroup>
 
 </Project>
