+++ tests/CreateTests.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Nodes;
 
-namespace Tests.ApiManifest;
+namespace Microsoft.OpenApi.ApiManifest.Tests;
 
 public class CreateTests
 {
@@ -52,8 +52,10 @@ public class CreateTests
     {
         _ = Assert.Throws<ArgumentException>(() =>
         {
-            var apiDependency = new ApiDependency();
-            apiDependency.ApiDeploymentBaseUrl = apiDeploymentBaseUrl;
+            var apiDependency = new ApiDependency
+            {
+                ApiDeploymentBaseUrl = apiDeploymentBaseUrl
+            };
         }
         );
     }
